# LUCENE-9368: SIMD-based decoding of BKD docIds