--- a/lucene/core/src/java/org/apache/lucene/util/bkd/BKDReader.java
+++ b/lucene/core/src/java/org/apache/lucene/util/bkd/BKDReader.java
@@ -484,6 +484,8 @@ public final class BKDReader extends PointValues implements Accountable {
     final IntersectVisitor visitor;
     public final IndexTree index;
 
+    final ForDocIdsWriter decoder = new ForDocIdsWriter();
+
     public IntersectState(IndexInput in, int numDims,
                           int packedBytesLength,
                           int packedIndexBytesLength,
@@ -527,7 +529,7 @@ public final class BKDReader extends PointValues implements Accountable {
       assert grown;
       //System.out.println("ADDALL");
       if (state.index.nodeExists()) {
-        visitDocIDs(state.in, state.index.getLeafBlockFP(), state.visitor);
+        visitDocIDs(state.in, state.index.getLeafBlockFP(), state.visitor, state.decoder);
       }
       // TODO: we can assert that the first value here in fact matches what the index claimed?
     } else {
@@ -556,30 +558,29 @@ public final class BKDReader extends PointValues implements Accountable {
   public void visitLeafBlockValues(IndexTree index, IntersectState state) throws IOException {
 
     // Leaf node; scan and filter all points in this block:
-    int count = readDocIDs(state.in, index.getLeafBlockFP(), state.scratchIterator);
+    int count = readDocIDs(state.in, index.getLeafBlockFP(), state.scratchIterator, state.decoder);
 
     // Again, this time reading values and checking with the visitor
     visitDocValues(state.commonPrefixLengths, state.scratchDataPackedValue, state.scratchMinIndexPackedValue, state.scratchMaxIndexPackedValue, state.in, state.scratchIterator, count, state.visitor);
   }
 
-  private void visitDocIDs(IndexInput in, long blockFP, IntersectVisitor visitor) throws IOException {
+  private void visitDocIDs(IndexInput in, long blockFP, IntersectVisitor visitor, ForDocIdsWriter decoder) throws IOException {
     // Leaf node
     in.seek(blockFP);
 
     // How many points are stored in this leaf cell:
     int count = in.readVInt();
     // No need to call grow(), it has been called up-front
-
-    DocIdsWriter.readInts(in, count, visitor);
+    DocIdsWriter.readInts(in, count, visitor, decoder);
   }
 
-  int readDocIDs(IndexInput in, long blockFP, BKDReaderDocIDSetIterator iterator) throws IOException {
+  int readDocIDs(IndexInput in, long blockFP, BKDReaderDocIDSetIterator iterator, ForDocIdsWriter decoder) throws IOException {
     in.seek(blockFP);
 
     // How many points are stored in this leaf cell:
     int count = in.readVInt();
 
-    DocIdsWriter.readInts(in, count, iterator.docIDs);
+    DocIdsWriter.readInts(in, count, iterator.docIDs, decoder);
 
     return count;
   }
@@ -784,7 +785,7 @@ public final class BKDReader extends PointValues implements Accountable {
       // In the unbalanced case it's possible the left most node only has one child:
       if (state.index.nodeExists()) {
         // Leaf node; scan and filter all points in this block:
-        int count = readDocIDs(state.in, state.index.getLeafBlockFP(), state.scratchIterator);
+        int count = readDocIDs(state.in, state.index.getLeafBlockFP(), state.scratchIterator, state.decoder);
 
         // Again, this time reading values and checking with the visitor
         visitDocValues(state.commonPrefixLengths, state.scratchDataPackedValue, state.scratchMinIndexPackedValue, state.scratchMaxIndexPackedValue, state.in, state.scratchIterator, count, state.visitor);