… lucene/core/src/java/org/apache/lucene/codecs/lucene84/ForUtil.java
… lucene/core/src/java/org/apache/lucene/util/bkd/ForDocIdsWriter.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene84/TestPForUtil.java
… lucene/core/src/test/org/apache/lucene/util/bkd/TestDocIdsWriter.java