This commit reuse the decoding technique used for posting to encode BKD docIds. One if the biggest differences is that in this case we want to expand our integers directly into an integer arrays, or feed the decoding values directly to the `IntersectVisitor`.

In order to do that we reorder the incoming docIds arrays so we can decode them later on in order.