@iverase 2020-05-27
I am going to close this in favour of https://github.com/apache/lucene-solr/pull/1538 as it less intrusive and performance test shows even better behaviour.